Extract cccc-lisp-kit; bundle Common Lisp + Emacs Lisp into cccc-lisp
The Lisp-family adapters shared ~80% of their code (collector stack,
walk_regions code-vs-data traversal, like-operator logical folding).
Hoist that skeleton into a new cccc-lisp-kit crate; a dialect is now
just a reader preset plus a head-symbol dispatch table (a LowerListFn +
LogicalOpFn pair).

- cccc-lisp-kit (new): shared Builder infra + lower/analyze drivers.
- cccc-scheme, cccc-clojure: rewritten as thin adapters over the kit;
  kept standalone (Scheme carries sub-dialects, Clojure has demand and
  a non-Lisp name). Behavior unchanged.
- cccc-lisp (new): bundles Common Lisp + Emacs Lisp, the traditional
  dialects not warranting their own crate. Its Dialect API also analyzes
  Scheme/Clojure by delegating to their crates rather than duplicating
  lowering.
- Remove cccc-commonlisp and cccc-emacslisp; rewire the CLI registry to
  cccc-lisp's commonlisp/emacslisp modules (--lang names unchanged).

Behavior-preserving: all existing tests pass unchanged and real-world
audit numbers are identical (chibi 7672, gauche 12486/3err, cl-ppcre
378, magit 1709, clojure 2582).
