+++ Cargo.toml
@@ -8,9 +8,9 @@ members = [
     "crates/cccc-php",
     "crates/cccc-rb",
     "crates/cccc-scheme",
-    "crates/cccc-commonlisp",
-    "crates/cccc-emacslisp",
     "crates/cccc-clojure",
+    "crates/cccc-lisp-kit",
+    "crates/cccc-lisp",
 ]
 resolver = "2"
 
@@ -30,9 +30,9 @@ cccc-go = { path = "crates/cccc-go", version = "1.1.0" }
 cccc-php = { path = "crates/cccc-php", version = "1.1.0" }
 cccc-rb = { path = "crates/cccc-rb", version = "1.1.0" }
 cccc-scheme = { path = "crates/cccc-scheme", version = "1.1.0" }
-cccc-commonlisp = { path = "crates/cccc-commonlisp", version = "1.1.0" }
-cccc-emacslisp = { path = "crates/cccc-emacslisp", version = "1.1.0" }
 cccc-clojure = { path = "crates/cccc-clojure", version = "1.1.0" }
+cccc-lisp-kit = { path = "crates/cccc-lisp-kit", version = "1.1.0" }
+cccc-lisp = { path = "crates/cccc-lisp", version = "1.1.0" }
 serde = { version = "1", features = ["derive"] }
 
 [profile.release]
