+++ crates/cccc-cli/Cargo.toml
@@ -20,8 +20,7 @@ cccc-go = { workspace = true }
 cccc-php = { workspace = true }
 cccc-rb = { workspace = true }
 cccc-scheme = { workspace = true }
-cccc-commonlisp = { workspace = true }
-cccc-emacslisp = { workspace = true }
+cccc-lisp = { workspace = true }
 cccc-clojure = { workspace = true }
 clap = { version = "4", features = ["derive"] }
 globset = "0.4"
