+++ crates/cccc-lisp/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "cccc-lisp"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Lisp-family adapter: Common Lisp, Emacs Lisp, and other Lisp dialects lowered into the cccc-core complexity IR"
+
+[dependencies]
+# The shared lowering kit (re-exports cccc-core IR + the lispexp reader).
+cccc-lisp-kit = { workspace = true }
+# Scheme and Clojure have their own standalone crates; cccc-lisp delegates to
+# them for those dialects (via `Dialect`), so their lowering lives in one place.
+cccc-scheme = { workspace = true }
+cccc-clojure = { workspace = true }
