+++ crates/cccc-scheme/Cargo.toml
@@ -9,6 +9,5 @@ homepage.workspace = true
 description = "Scheme (R7RS-small) adapter that lowers S-expressions into the cccc-core complexity IR"
 
 [dependencies]
-cccc-core = { workspace = true }
-# Pure-Rust, position-annotated, code-vs-data-aware S-expression reader.
-lispexp = "0.5"
+# The shared Lisp-family lowering kit (re-exports cccc-core IR + the lispexp reader).
+cccc-lisp-kit = { workspace = true }
